Allowed lukiodiplomi for aineopiskelija
