+++ framework/src/main/java/fi/otavanopisto/pyramus/koski/model/lukio/ops2019/LukionOppiaineenOppimaaranSuoritus2019.java
@@ -1,8 +1,5 @@
 package fi.otavanopisto.pyramus.koski.model.lukio.ops2019;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import com.fasterxml.jackson.databind.JsonDeserializer;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 
@@ -22,18 +19,9 @@ public class LukionOppiaineenOppimaaranSuoritus2019 extends LukionSuoritus2019 {
     this.koulutusmoduuli.setPerusteenDiaarinumero(perusteenDiaarinumero);
   }
   
-  public void addOsasuoritus(LukionOppiaineenSuoritus2019 osasuoritus) {
-    osasuoritukset.add(osasuoritus);
-  }
-  
-  public List<LukionOppiaineenSuoritus2019> getOsasuoritukset() {
-    return osasuoritukset;
-  }
-
   public LukionOppiaineenOppimaaranKoulutusmoduuli2019 getKoulutusmoduuli() {
     return koulutusmoduuli;
   }
 
   private final LukionOppiaineenOppimaaranKoulutusmoduuli2019 koulutusmoduuli = new LukionOppiaineenOppimaaranKoulutusmoduuli2019();
-  private final List<LukionOppiaineenSuoritus2019> osasuoritukset = new ArrayList<>();
 }
