+++ framework/src/main/java/fi/otavanopisto/pyramus/koski/model/lukio/ops2019/LukionOppimaaranSuoritus2019.java
@@ -1,8 +1,5 @@
 package fi.otavanopisto.pyramus.koski.model.lukio.ops2019;
 
-import java.util.ArrayList;
-import java.util.List;
-
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.databind.JsonDeserializer;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
@@ -25,19 +22,10 @@ public class LukionOppimaaranSuoritus2019 extends LukionSuoritus2019 {
       OrganisaationToimipiste toimipiste) {
     super(SuorituksenTyyppi.lukionoppimaara, oppimaara, suorituskieli, toimipiste);
   }
-  
-  public void addOsasuoritus(LukionOsasuoritus2019 osasuoritus) {
-    osasuoritukset.add(osasuoritus);
-  }
-  
-  public List<LukionOsasuoritus2019> getOsasuoritukset() {
-    return osasuoritukset;
-  }
-  
+    
   public Koulutusmoduuli getKoulutusmoduuli() {
     return koulutusmoduuli;
   }
 
   private final Koulutusmoduuli koulutusmoduuli = new Koulutusmoduuli(Koulutus.K309902);
-  private final List<LukionOsasuoritus2019> osasuoritukset = new ArrayList<>();
 }
